backend/requirements.txt
fastapi>=0.114.0
uvicorn>=0.30.0
sqlalchemy>=2.0.30
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
redis>=5.0.8
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
alembic>=1.13.1
google-auth>=2.29.0
requests>=2.31.0
fastapi-limiter>=0.1.6
